refactor: migrate aws_vpc_endpoint to new structure
